# src/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Weynon_V2",
    "description": "Welcome to Weynon Api",
    "contact": {}
  },
  "host": "178.79.177.139:8848/api",
  "basePath": "/",
  "securityDefinitions": {
    "auth": {
      "type": "oauth2",
      "flow": "implicit",
      "authorizationUrl": "http://178.79.177.139:8848/api",
      "scopes": {}
    }
  },
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/friends": {
      "get": {
        "description": "TODO: Add Description",
        "summary": "Get All Friends",
        "tags": [
          "Friends"
        ],
        "operationId": "FriendsGet",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          },
          {
            "name": "offset",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "auth": []
          }
        ]
      }
    },
    "/friends/invite": {
      "post": {
        "description": "TODO: Add Description",
        "summary": "Invite A User For Friendship",
        "tags": [
          "Friends"
        ],
        "operationId": "FriendsInvitePost",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/InviteAUserForFriendshipRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "auth": []
          }
        ]
      }
    },
    "/friends/friend": {
      "post": {
        "description": "TODO: Add Description",
        "summary": "Confirm Friend Request",
        "tags": [
          "Friends"
        ],
        "operationId": "FriendsFriendPost",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ConfirmFriendRequestRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "auth": []
          }
        ]
      },
      "delete": {
        "description": "TODO: Add Description",
        "summary": "Delete Friend Request/Remove User From Friend List",
        "tags": [
          "Friends"
        ],
        "operationId": "FriendsFriendDelete",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/DeleteFriendRequest~1RemoveUserFromFriendListRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "auth": []
          }
        ]
      }
    },
    "/friends/change-friend-group": {
      "patch": {
        "description": "TODO: Add Description",
        "summary": "Change Friend Group",
        "tags": [
          "Friends"
        ],
        "operationId": "FriendsChangeFriendGroupPatch",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ChangeFriendGroupRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "auth": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "InviteAUserForFriendshipRequest": {
      "title": "InviteAUserForFriendshipRequest",
      "example": {
        "groupId": "5c7d1eb44b80fa35c0bc776e",
        "userId": "5c5aae5ce2ecdbbc144ed302"
      },
      "type": "object",
      "properties": {
        "groupId": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        }
      },
      "required": [
        "groupId",
        "userId"
      ]
    },
    "ConfirmFriendRequestRequest": {
      "title": "ConfirmFriendRequestRequest",
      "example": {
        "groupId": "5c7d1eb54b80fa35c0bc776f",
        "userId": "5c7d1ed74be10635d449bb29"
      },
      "type": "object",
      "properties": {
        "groupId": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        }
      },
      "required": [
        "groupId",
        "userId"
      ]
    },
...
    "ChangeFriendGroupRequest": {
      "title": "ChangeFriendGroupRequest",
      "example": {
        "groupId": "5c7d1eb54b80fa35c0bc776f",
        "userId": "5c7d1ed74be10635d449bb29"
      },
      "type": "object",
      "properties": {
        "groupId": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        }
      },
      "required": [
        "groupId",
        "userId"
      ]
    }
  },
...
}